# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Produtos",
        "description": "Documentação da API para salvar categorias, locais e produtos",
        "version": "1.0.0",
        "contact": {
            "name": "Seu Nome",
            "email": "email@exemplo.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000"
        }
    ],
    "paths": {
...
        "/categories/add": {
            "post": {
                "summary": "Cria uma nova categoria",
                "tags": [
                    "Categorias"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Category"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoria criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Category"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao criar a categoria"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Category": {
                "type": "object",
                "required": [
                    "nome"
                ],
                "properties": {
                    "nome": {
                        "type": "string",
                        "description": "Nome da categoria"
                    }
                },
                "example": {
                    "nome": "Eletrônicos"
                }
            },
...
        }
    }
}